docs: improved documentation, separated redis related things from cache
